+++ packages/client/src/routes/Home/mapUtils.ts
@@ -2,7 +2,7 @@ import moment from 'moment'
 
 import { PolylineConfig } from '../../components/Map/Component'
 import { decodePath } from '../../components/Map/utils'
-import { Entry } from '../../services/timeline'
+import { Entry } from '../../services/timeline/types'
 
 const mapElementColors = {
   default: '#eb4434',
